Move commands.py to telemetry module
